+++ ttyd-tools/rel/source/tot_manager_options.cpp
@@ -185,14 +185,12 @@ void OptionsManager::InitFromSelectedOptions() {
             int32_t num_items = state.GetOption(STAT_PERM_ITEM_LOAD_SIZE);
             int32_t num_badges = state.GetOption(STAT_PERM_BADGE_LOAD_SIZE);
             for (int32_t i = num_items - 1; i >= 0; --i) {
-                int32_t item_id =
-                    static_cast<uint8_t>(state.GetOption(STAT_PERM_ITEM_LOADOUT, i))
+                int32_t item_id = state.GetOption(STAT_PERM_ITEM_LOADOUT, i)
                     + ItemType::THUNDER_BOLT;
                 ttyd::mario_pouch::pouchGetItem(item_id);
             }
             for (int32_t i = num_badges - 1; i >= 0; --i) {
-                int32_t item_id =
-                    static_cast<uint8_t>(state.GetOption(STAT_PERM_BADGE_LOADOUT, i))
+                int32_t item_id = state.GetOption(STAT_PERM_BADGE_LOADOUT, i)
                     + ItemType::THUNDER_BOLT;
                 ttyd::mario_pouch::pouchGetItem(item_id);
             }
